requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==4.2.7
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-taggit==6.0.0
django-widget-tweaks==1.5.0
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.7.0
